14/3/24 today work on landingpage hoast landing page and work on admin panel of apps and more
